+++ app/src/main/java/com/strukovnasamobor/samobornt/GeofenceReceiver.kt
@@ -8,7 +8,6 @@ import com.google.android.gms.location.Geofence
 import com.google.android.gms.location.GeofencingEvent
 
 class GeofenceReceiver : BroadcastReceiver() {
-    lateinit var key: String
     override fun onReceive(context: Context?, intent: Intent?) {
         if (context != null) {
             val geofencingEvent = GeofencingEvent.fromIntent(intent)
@@ -18,6 +17,11 @@ class GeofenceReceiver : BroadcastReceiver() {
 
                 triggeringGeofences.forEach {
                     Toast.makeText(context, "Ulaz u geofence: " + it.requestId, Toast.LENGTH_SHORT).show()
+
+                    MapboxActivity.showNotification(
+                        context.applicationContext,
+                        it.requestId
+                    )
                     //ovdje umjesto ovog Toast-a, treba poslati notifikaciju da je korisnik usao u geofence, ime geofence-a dobijemo preko: it.requestId
                 }
