+++ app/src/main/java/com/strukovnasamobor/samobornt/MapboxActivity.kt
@@ -1,9 +1,19 @@
 package com.strukovnasamobor.samobornt
 
+import android.Manifest
 import android.annotation.SuppressLint
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.database.Cursor
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
+import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
 import com.mapbox.android.core.permissions.PermissionsListener
 import com.mapbox.android.core.permissions.PermissionsManager
@@ -18,12 +28,25 @@ import com.mapbox.mapboxsdk.maps.MapView
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback
 import com.mapbox.mapboxsdk.maps.Style
+import com.strukovnasamobor.samobornt.services.*
+import kotlin.random.Random
+import androidx.core.app.ActivityCompat
+import com.google.android.gms.location.Geofence
+import com.google.android.gms.location.GeofencingClient
+import com.google.android.gms.location.GeofencingRequest
+import com.google.android.gms.location.LocationServices
+import com.strukovnasamobor.samobornt.detail.DetailActivity
+
+const val GEOFENCE_RADIUS = 100 //ovo je radius geofence-a, tj. koliko daleko korisnik mora biti da dobije notifikaciju
+const val GEOFENCE_LOCATION_REQUEST_CODE = 5
 
 class MapboxActivity : BaseActivity(), OnMapReadyCallback, PermissionsListener {
 
     private var permissionsManager: PermissionsManager = PermissionsManager(this)
     private lateinit var mapboxMap: MapboxMap
     private var mapView: MapView? = null
+    private lateinit var connection: DBConnection
+    private lateinit var geofencingClient: GeofencingClient
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -31,14 +54,41 @@ class MapboxActivity : BaseActivity(), OnMapReadyCallback, PermissionsListener {
         // object or in the same activity which contains the mapview.
         Mapbox.getInstance(this, getString(R.string.mapbox_access_token))
 
+        connection = DBConnection.getConnectionInstance(this)
+        geofencingClient = LocationServices.getGeofencingClient(this)
+
         // This contains the MapView in XML and needs to be called after the access token is configured.
         setContentView(R.layout.activity_mapbox)
         super.initializeMenu()
 
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            if (ContextCompat.checkSelfPermission(
+                    applicationContext,
+                    Manifest.permission.ACCESS_BACKGROUND_LOCATION
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                ActivityCompat.requestPermissions(
+                    this,
+                    arrayOf(Manifest.permission.ACCESS_BACKGROUND_LOCATION),
+                    GEOFENCE_LOCATION_REQUEST_CODE
+                )
+            }
+        }
+
+        val geofenceLocations = getLocationList()
+        geofenceLocations.forEach {
+            createGeofence(
+                LatLng(
+                    it.second,
+                    it.third
+                ), it.first!!, geofencingClient
+            )
+        }
+
         mapView = findViewById(R.id.mapView);
         mapView?.onCreate(savedInstanceState)
         mapView?.getMapAsync(this)
-    }
+        }
 
     override fun onMapReady(mapboxMap: MapboxMap) {
         this.mapboxMap = mapboxMap
@@ -191,4 +241,105 @@ class MapboxActivity : BaseActivity(), OnMapReadyCallback, PermissionsListener {
         super.onLowMemory()
         mapView?.onLowMemory()
     }
+
+    private fun createGeofence(location: LatLng, key: String, geofencingClient: GeofencingClient) {
+        val geofence = Geofence.Builder()
+            .setRequestId(key)
+            .setCircularRegion(location.latitude, location.longitude, GEOFENCE_RADIUS.toFloat())
+            .setExpirationDuration(Geofence.NEVER_EXPIRE)
+            .setTransitionTypes(Geofence.GEOFENCE_TRANSITION_ENTER)
+            .build()
+
+        val geofenceRequest = GeofencingRequest.Builder()
+            .setInitialTrigger(GeofencingRequest.INITIAL_TRIGGER_ENTER)
+            .addGeofence(geofence)
+            .build()
+
+        val intent = Intent(this, GeofenceReceiver::class.java)
+
+        val pendingIntent = PendingIntent.getBroadcast(
+            applicationContext,
+            0,
+            intent,
+            PendingIntent.FLAG_UPDATE_CURRENT
+        )
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            if (ContextCompat.checkSelfPermission(
+                    applicationContext,
+                    Manifest.permission.ACCESS_BACKGROUND_LOCATION
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
+                ActivityCompat.requestPermissions(
+                    this,
+                    arrayOf(Manifest.permission.ACCESS_BACKGROUND_LOCATION),
+                    GEOFENCE_LOCATION_REQUEST_CODE
+                )
+            } else {
+                geofencingClient.addGeofences(geofenceRequest, pendingIntent)
+            }
+        } else {
+            geofencingClient.addGeofences(geofenceRequest, pendingIntent)
+        }
+    }
+
+    companion object {
+        fun showNotification(context: Context?, message: String) {
+            val CHANNEL_ID = "REMINDER_NOTIFICATION_CHANNEL"
+            var notificationId = 123
+            notificationId += Random(notificationId).nextInt(1, 30)
+
+            val notificationString = "Došli ste do: $message, kliknite ovdje da pročitate više!"
+
+            val resultIntent = Intent(context, DetailActivity::class.java)
+                .putExtra("locationName", message)
+
+            val resultPendingIntent = PendingIntent.getActivity(context, 1, resultIntent, PendingIntent.FLAG_UPDATE_CURRENT)
+
+            val notificationBuilder = NotificationCompat.Builder(context!!.applicationContext, CHANNEL_ID)
+                .setSmallIcon(R.drawable.marker_custom)
+                .setContentTitle(context.getString(R.string.app_name))
+                .setContentText(notificationString)
+                .setAutoCancel(true)
+                .setContentIntent(resultPendingIntent)
+                .setStyle(
+                    NotificationCompat.BigTextStyle()
+                        .bigText(notificationString)
+                )
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+
+            val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                val channel = NotificationChannel(
+                    CHANNEL_ID,
+                    context.getString(R.string.app_name),
+                    NotificationManager.IMPORTANCE_DEFAULT
+                ).apply {
+                    description = context.getString(R.string.app_name)
+                }
+                notificationManager.createNotificationChannel(channel)
+            }
+            //Toast.makeText(context, "Sending notification!", Toast.LENGTH_SHORT).show()
+            notificationManager.notify(notificationId, notificationBuilder.build())
+        }
+    }
+
+    private fun getLocationList(): MutableList<Triple<String, Double, Double>> {
+        val locationList: MutableList<Triple<String, Double, Double>> = mutableListOf()
+        val cardsListLocale = resources.configuration.locales[0].toString()
+        val correctTable = if (cardsListLocale == "hr") TABLE_NAME_HRV else TABLE_NAME_ENG
+        val cursor: Cursor = connection.getFetchAllCursor(correctTable)
+        if (cursor.moveToFirst()) {
+            while (!cursor.isAfterLast) {
+                val locationName = cursor.getString(cursor.getColumnIndex(C_NAME))
+                val latitude = cursor.getDouble(cursor.getColumnIndex(C_LATITUDE))
+                val longitude = cursor.getDouble(cursor.getColumnIndex(C_LONGITUDE))
+                locationList.add(Triple(locationName, latitude, longitude))
+                cursor.moveToNext()
+            }
+        }
+        return locationList
+    }
+
 }
